Amended controller actions.

+++ WebbiSkools.QuizManager.Web/Controllers/QuizzesController.cs
@@ -122,7 +122,7 @@ namespace WebbiSkools.QuizManager.Web.Controllers
             return View(quiz);
         }
 
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -131,9 +131,15 @@ namespace WebbiSkools.QuizManager.Web.Controllers
 
             var quiz = await _context.Quizzes
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (quiz == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] = "Delete failed. Please try again.";
             }
 
             return View(quiz);
@@ -144,9 +150,22 @@ namespace WebbiSkools.QuizManager.Web.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var quiz = await _context.Quizzes.FindAsync(id);
-            _context.Quizzes.Remove(quiz);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if (quiz == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Quizzes.Remove(quiz);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch
+            {
+                return RedirectToAction(nameof(Delete), new {id = id, saveChnagesError = true});
+            }
         }
 
         private bool QuizExists(int id)
